Rename shoot upgrade step and steps' files
