+++ components/provisioner/internal/operations/stages/shootupgrade/wait_for_shoot_new_version_test.go
@@ -5,21 +5,31 @@ import (
 	"testing"
 	"time"
 
+	gardencorev1beta1 "github.com/gardener/gardener/pkg/apis/core/v1beta1"
+	gardener_types "github.com/gardener/gardener/pkg/apis/core/v1beta1"
+	gardener_mocks "github.com/kyma-project/control-plane/components/provisioner/internal/operations/stages/deprovisioning/mocks"
+
 	"github.com/kyma-project/control-plane/components/provisioner/internal/model"
 	"github.com/kyma-project/control-plane/components/provisioner/internal/operations"
-	gardener_mocks "github.com/kyma-project/control-plane/components/provisioner/internal/operations/stages/deprovisioning/mocks"
 	"github.com/kyma-project/control-plane/components/provisioner/internal/operations/testkit"
+
 	"github.com/sirupsen/logrus"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/mock"
 	"github.com/stretchr/testify/require"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 )
 
-func TestWaitForClusterUpgrade(t *testing.T) {
+const (
+	oldGeneration = 1
+	newGeneration = 2
+)
 
+func TestWaitForNewShootVersion(t *testing.T) {
 	clusterName := "shootName"
 	runtimeID := "runtimeID"
 	tenant := "tenant"
+	operationID := "operationID"
 
 	cluster := model.Cluster{
 		ID:     runtimeID,
@@ -36,59 +46,27 @@ func TestWaitForClusterUpgrade(t *testing.T) {
 		expectedDelay time.Duration
 	}{
 		{
-			description: "should continue waiting if cluster is in processing state",
+			description: "should continue waiting if cluster has old resource version",
 			mockFunc: func(gardenerClient *gardener_mocks.GardenerClient) {
 				gardenerClient.On("Get", clusterName, mock.Anything).Return(
 					testkit.NewTestShoot(clusterName).
-						WithOperationProcessing().
+						WithGeneration(2).
+						WithObservedGeneration(1).
+						ToShoot(), nil)
+			},
+			expectedStage: model.WaitingForShootNewVersion,
+			expectedDelay: 5 * time.Second,
+		},
+		{
+			description: "should move to next step if resource version changes",
+			mockFunc: func(gardenerClient *gardener_mocks.GardenerClient) {
+				gardenerClient.On("Get", clusterName, mock.Anything).Return(
+					testkit.NewTestShoot(clusterName).
+						WithGeneration(2).
+						WithObservedGeneration(2).
 						ToShoot(), nil)
 			},
 			expectedStage: model.WaitingForShootUpgrade,
-			expectedDelay: 20 * time.Second,
-		},
-		{
-			description: "should continue waiting if cluster is in pending state",
-			mockFunc: func(gardenerClient *gardener_mocks.GardenerClient) {
-				gardenerClient.On("Get", clusterName, mock.Anything).Return(
-					testkit.NewTestShoot(clusterName).
-						WithOperationPending().
-						ToShoot(), nil)
-			},
-			expectedStage: model.WaitingForShootUpgrade,
-			expectedDelay: 20 * time.Second,
-		},
-		{
-			description: "should continue waiting if cluster is in error state - the operation will be retried on Gardener side",
-			mockFunc: func(gardenerClient *gardener_mocks.GardenerClient) {
-				gardenerClient.On("Get", clusterName, mock.Anything).Return(
-					testkit.NewTestShoot(clusterName).
-						WithOperationError().
-						ToShoot(), nil)
-			},
-			expectedStage: model.WaitingForShootUpgrade,
-			expectedDelay: 20 * time.Second,
-		},
-		{
-			description: "should continue waiting if last operation not set",
-			mockFunc: func(gardenerClient *gardener_mocks.GardenerClient) {
-				gardenerClient.On("Get", clusterName, mock.Anything).Return(
-					testkit.NewTestShoot(clusterName).
-						WithOperationNil().
-						ToShoot(), nil)
-			},
-			expectedStage: model.WaitingForShootUpgrade,
-			expectedDelay: 20 * time.Second,
-		},
-		{
-			description: "should return finished stage if cluster upgrade has succeeded",
-			mockFunc: func(gardenerClient *gardener_mocks.GardenerClient) {
-
-				gardenerClient.On("Get", clusterName, mock.Anything).Return(
-					testkit.NewTestShoot(clusterName).
-						WithOperationSucceeded().
-						ToShoot(), nil)
-			},
-			expectedStage: model.FinishedStage,
 			expectedDelay: 0,
 		},
 	} {
@@ -98,9 +76,10 @@ func TestWaitForClusterUpgrade(t *testing.T) {
 
 			testCase.mockFunc(gardenerClient)
 
-			waitForrShootClusterUpgradeStep := NewWaitForShootClusterUpgradeStep(gardenerClient, model.FinishedStage, time.Minute)
+			waitForShootClusterUpgradeStep := NewWaitForShootNewVersionStep(gardenerClient, model.WaitingForShootUpgrade, time.Minute)
+
 			// when
-			result, err := waitForrShootClusterUpgradeStep.Run(cluster, model.Operation{}, logrus.New())
+			result, err := waitForShootClusterUpgradeStep.Run(cluster, model.Operation{ID: operationID}, logrus.New())
 
 			// then
 			require.NoError(t, err)
@@ -129,6 +108,8 @@ func TestWaitForClusterUpgrade(t *testing.T) {
 			mockFunc: func(gardenerClient *gardener_mocks.GardenerClient) {
 				gardenerClient.On("Get", clusterName, mock.Anything).Return(
 					testkit.NewTestShoot(clusterName).
+						WithGeneration(2).
+						WithObservedGeneration(1).
 						WithOperationFailed().
 						ToShoot(), nil)
 			},
@@ -142,16 +123,58 @@ func TestWaitForClusterUpgrade(t *testing.T) {
 
 			testCase.mockFunc(gardenerClient)
 
-			waitForClusterCreationStep := NewWaitForShootClusterUpgradeStep(gardenerClient, model.FinishedStage, time.Minute)
+			waitForClusterCreationStep := NewWaitForShootNewVersionStep(gardenerClient, model.FinishedStage, time.Minute)
 
 			// when
 			_, err := waitForClusterCreationStep.Run(testCase.cluster, model.Operation{}, logrus.New())
 
 			// then
 			require.Error(t, err)
-			nonRecoverable := operations.NonRecoverableError{}
-			require.Equal(t, testCase.unrecoverableError, errors.As(err, &nonRecoverable))
+			require.Equal(t, testCase.unrecoverableError, errors.As(err, &operations.NonRecoverableError{}))
 			gardenerClient.AssertExpectations(t)
 		})
 	}
+}
+
+type testShoot struct {
+	shoot *gardener_types.Shoot
+}
+
+func newTestShoot(name string) *testShoot {
+	return &testShoot{
+		shoot: &gardener_types.Shoot{
+			TypeMeta: metav1.TypeMeta{},
+			ObjectMeta: metav1.ObjectMeta{
+				Name: name,
+			},
+			Spec: gardener_types.ShootSpec{},
+			Status: gardener_types.ShootStatus{
+				LastOperation: &gardener_types.LastOperation{},
+			},
+		},
+	}
+}
+
+func (ts *testShoot) toShoot() *gardener_types.Shoot {
+	return ts.shoot
+}
+
+func (ts *testShoot) withGeneration(generation int64) *testShoot {
+	ts.shoot.Generation = generation
+	return ts
+}
+
+func (ts *testShoot) withObservedGeneration(generation int64) *testShoot {
+	ts.shoot.Status.ObservedGeneration = generation
+	return ts
+}
+
+func (ts *testShoot) withOperationSucceeded() *testShoot {
+	ts.shoot.Status.LastOperation.State = gardencorev1beta1.LastOperationStateSucceeded
+	return ts
+}
+
+func (ts *testShoot) withOperationFailed() *testShoot {
+	ts.shoot.Status.LastOperation.State = gardencorev1beta1.LastOperationStateFailed
+	return ts
 }
